refactor: move `create_or_update_smart_contract/2` to `SmartContract` module
