+++ apps/explorer/lib/explorer/smart_contract/solidity/publisher.ex
@@ -293,51 +293,15 @@ defmodule Explorer.SmartContract.Solidity.Publisher do
   def publish_smart_contract(address_hash, params, abi) do
     attrs = address_hash |> attributes(params, abi)
 
-    create_or_update_smart_contract(address_hash, attrs)
+    Logger.info("Publish successfully verified Solidity smart-contract #{address_hash} into the DB")
+    SmartContract.create_or_update_smart_contract(address_hash, attrs)
   end
 
   def publish_smart_contract(address_hash, params, abi, file_path) do
     attrs = address_hash |> attributes(params, file_path, abi)
 
-    create_or_update_smart_contract(address_hash, attrs)
-  end
-
-  @doc """
-    Creates or updates a smart contract record based on its verification status.
-
-    This function first checks if a smart contract associated with the provided address hash
-    is already verified. If verified, it updates the existing smart contract record with the
-    new attributes provided, such as external libraries and secondary sources. During the update,
-    the contract methods are also updated: existing methods are preserved, and any new methods
-    from the provided ABI are added to ensure the contract's integrity and completeness.
-
-    If the smart contract is not verified, it creates a new record in the database with the
-    provided attributes, setting it up for verification. In this case, all contract methods
-    from the ABI are freshly inserted as part of the new smart contract creation.
-
-    ## Parameters
-    - `address_hash`: The hash of the address for the smart contract.
-    - `attrs`: A map containing attributes such as external libraries and secondary sources.
-
-    ## Returns
-    - `{:ok, Explorer.Chain.SmartContract.t()}`: Successfully created or updated smart
-      contract.
-    - `{:error, data}`: on failure, returning `Ecto.Changeset.t()` or, if any issues
-      happen during setting the address as verified, an error message.
-  """
-  @spec create_or_update_smart_contract(binary() | Explorer.Chain.Hash.t(), %{
-          :external_libraries => list(),
-          :secondary_sources => list(),
-          optional(any()) => any()
-        }) :: {:error, Ecto.Changeset.t() | String.t()} | {:ok, Explorer.Chain.SmartContract.t()}
-  def create_or_update_smart_contract(address_hash, attrs) do
     Logger.info("Publish successfully verified Solidity smart-contract #{address_hash} into the DB")
-
-    if SmartContract.verified?(address_hash) do
-      SmartContract.update_smart_contract(attrs, attrs.external_libraries, attrs.secondary_sources)
-    else
-      SmartContract.create_smart_contract(attrs, attrs.external_libraries, attrs.secondary_sources)
-    end
+    SmartContract.create_or_update_smart_contract(address_hash, attrs)
   end
 
   defp unverified_smart_contract(address_hash, params, error, error_message, verification_with_files? \\ false) do
